# [1.20.x] More buildscript clean-up