--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/DownloadLibraries.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/DownloadLibraries.groovy
@@ -27,7 +27,7 @@ abstract class DownloadLibraries extends DefaultTask {
     def run() {
 		Util.init()
 		File outputDir = output.get().asFile
-        def libraries = new ArrayList()
+        var libraries = new ArrayList<String>()
 
 		def json = input.get().asFile.json().libraries.each { lib ->
 		    //TODO: Thread?
@@ -35,7 +35,7 @@ abstract class DownloadLibraries extends DefaultTask {
 			artifacts.each{ art -> 
 				def target = new File(outputDir, art.path)
 				libraries.add(target.absolutePath)
-				if (!target.exists() || !art.sha1.equals(target.sha1())) {
+				if (!target.exists() || art.sha1 != target.sha1()) {
 					project.logger.lifecycle("Downloading ${art.url}")
 					if (!target.parentFile.exists()) {
 						target.parentFile.mkdirs()
@@ -43,7 +43,7 @@ abstract class DownloadLibraries extends DefaultTask {
 					new URL(art.url).withInputStream { i ->
 						target.withOutputStream { it << i }
 					}
-					if (!art.sha1.equals(target.sha1())) {
+					if (art.sha1 != target.sha1()) {
 						throw new IllegalStateException("Failed to download ${art.url} to ${target.canonicalPath} SHA Mismatch")
 					}
 				}