--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/ExtractFile.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/ExtractFile.groovy
@@ -1,10 +1,14 @@
 package net.minecraftforge.forge.tasks
 
+import groovy.transform.CompileStatic
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.*
 
+import java.util.zip.ZipFile
+
+@CompileStatic
 abstract class ExtractFile extends DefaultTask {
     @InputFile abstract RegularFileProperty getInput()
     @Input abstract Property<String> getTarget()
@@ -12,11 +16,11 @@ abstract class ExtractFile extends DefaultTask {
 
     @TaskAction
     protected void exec() {
-        def zip = new java.util.zip.ZipFile(input.get().asFile);
-        def entry = zip.getEntry(target.get())
+        var zip = new ZipFile(input.get().asFile);
+        var entry = zip.getEntry(target.get())
         if (entry == null)
             throw new IllegalStateException(input.get().asFile.absolutePath + " does not contain " + target.get())
-        def stream = zip.getInputStream(entry)
+        var stream = zip.getInputStream(entry)
         output.get().asFile.withOutputStream { out ->
             stream.transferTo(out)
         }