--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/FieldCompareFinder.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/FieldCompareFinder.groovy
@@ -1,20 +1,11 @@
 package net.minecraftforge.forge.tasks
 
+import groovy.transform.CompileStatic
 import groovy.transform.EqualsAndHashCode
 
-import java.util.ArrayList
-import java.util.HashMap
-import java.util.TreeMap
-import java.util.TreeSet
-import java.util.function.BiConsumer
-
-import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.*
-
-import org.objectweb.asm.Opcodes
-import org.objectweb.asm.Type
+import org.objectweb.asm.tree.AbstractInsnNode
 import org.objectweb.asm.tree.ClassNode
-import org.objectweb.asm.tree.FieldNode
 import org.objectweb.asm.tree.MethodNode
 
 import static org.objectweb.asm.Opcodes.*
@@ -34,17 +25,17 @@ abstract class FieldCompareFinder extends BytecodeFinder {
     
     @Override
     protected process(ClassNode parent, MethodNode node) {
-        def last = null
+        AbstractInsnNode last = null
         def parentInstance = new ObjectTarget(owner: parent.name, name: '', desc: '')
         for (int x = 0; x < node.instructions.size(); x++) {
             def current = node.instructions.get(x)
-            if (current.opcode == IF_ACMPEQ || current.opcode == IF_ACMPNE) {
-                if (last != null && (last.opcode == GETSTATIC || last.opcode == GETFIELD)) {
+            if (current.opcode === IF_ACMPEQ || current.opcode === IF_ACMPNE) {
+                if (last !== null && (last.opcode === GETSTATIC || last.opcode === GETFIELD)) {
                     def target = new Search(cls: last.owner, name: last.name)
                     def wanted = fieldsReverse.get(target)
                     def original = fields.get(wanted)
                     def instance = new ObjectTarget(owner: parent.name, name: node.name, desc: node.desc)
-                    if (wanted != null && (original.blacklist == null || (!original.blacklist.contains(instance) && !original.blacklist.contains(parentInstance)))) {
+                    if (wanted !== null && (original.blacklist === null || (!original.blacklist.contains(instance) && !original.blacklist.contains(parentInstance)))) {
                         targets.computeIfAbsent(wanted, { k -> new TreeSet() }).add(instance)
                     }
                 }
@@ -68,9 +59,10 @@ abstract class FieldCompareFinder extends BytecodeFinder {
 		}
         return ret
     }
-    
+
+    @CompileStatic
     @EqualsAndHashCode(excludes = ['replacement', 'blacklist'])
-    public static class Search {
+    static class Search {
         @Input
         String cls
         
@@ -89,17 +81,17 @@ abstract class FieldCompareFinder extends BytecodeFinder {
             return cls + '.' + name
         }
         
-        def blacklist(def owner, def name, def desc) {
-            if (blacklist == null)
-                blacklist = new HashSet()
+        def blacklist(String owner, String name, String desc) {
+            if (blacklist === null)
+                blacklist = new HashSet<>()
             blacklist.add(new ObjectTarget(owner: owner, name: name, desc: desc))
         }
-        def blacklist(def owner) {
+        def blacklist(String owner) {
             blacklist(owner, '', '')
         }
     }
     
-    def fields(Closure cl) {
+    void fields(Closure cl) {
         new ClosureHelper(cl, {name, ccl ->
             def search = ClosureHelper.apply(new Search(), ccl)
             this.fields.put(name, search)