--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/InstallerJson.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/InstallerJson.groovy
@@ -7,6 +7,7 @@ import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.*
+import org.gradle.api.tasks.bundling.AbstractArchiveTask
 
 import java.nio.file.Files
 
@@ -45,7 +46,7 @@ abstract class InstallerJson extends DefaultTask {
         [
             project.tasks.universalJar,
             project.tasks.serverShimJar
-        ].forEach { packed ->
+        ].forEach { AbstractArchiveTask packed ->
             def info = Util.getMavenInfoFromTask(packed)
             libs.put(info.name, [
                 name: info.name,