--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/LauncherJson.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/LauncherJson.groovy
@@ -6,6 +6,7 @@ import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.*
+import org.gradle.api.tasks.bundling.AbstractArchiveTask
 
 import java.nio.file.Files
 
@@ -58,38 +59,35 @@ abstract class LauncherJson extends DefaultTask {
 
     @TaskAction
     protected void exec() {
-        [
-            project.tasks.universalJar
-        ].forEach { packed ->
-            def info = Util.getMavenInfoFromTask(packed)
-            json.libraries.add([
-                name: info.name,
-                downloads: [
-                    artifact: [
-                        path: info.path,
-                        url: "https://maven.minecraftforge.net/$info.path",
-                        sha1: packed.archiveFile.get().asFile.sha1(),
-                        size: packed.archiveFile.get().asFile.length()
-                    ]
+        var packed = (AbstractArchiveTask) project.tasks.universalJar
+        def info = Util.getMavenInfoFromTask(packed)
+        json.libraries.add([
+            name: info.name,
+            downloads: [
+                artifact: [
+                    path: info.path,
+                    url: "https://maven.minecraftforge.net/$info.path",
+                    sha1: packed.archiveFile.get().asFile.sha1(),
+                    size: packed.archiveFile.get().asFile.length()
                 ]
-            ])
-        }
-        [
-            'client': project.tasks.applyClientBinPatches
-        ].forEach { classifier, genned ->
-            def info = Util.getMavenInfoFromTask(genned, classifier)
-            json.libraries.add([
-                name: info.name,
-                downloads: [
-                    artifact: [
-                        path: info.path,
-                        url: "",
-                        sha1: genned.output.get().asFile.sha1(),
-                        size: genned.output.get().asFile.length()
-                    ]
+            ]
+        ])
+
+        var classifier = 'client'
+        var genned = project.tasks.applyClientBinPatches
+        info = Util.getMavenInfoFromTask(genned, classifier)
+        json.libraries.add([
+            name: info.name,
+            downloads: [
+                artifact: [
+                    path: info.path,
+                    url: "",
+                    sha1: genned.output.get().asFile.sha1(),
+                    size: genned.output.get().asFile.length()
                 ]
-            ])
-        }
+            ]
+        ])
+
         getArtifacts(project, project.configurations.installer).each { key, lib -> 
             json.libraries.add(lib)
         }