--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/MergeJars.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/MergeJars.groovy
@@ -1,5 +1,6 @@
 package net.minecraftforge.forge.tasks
 
+import groovy.transform.CompileStatic
 import org.apache.commons.io.IOUtils
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.ConfigurableFileCollection
@@ -12,6 +13,7 @@ import java.util.zip.ZipEntry
 import java.util.zip.ZipInputStream
 import java.util.zip.ZipOutputStream
 
+@CompileStatic
 abstract class MergeJars extends DefaultTask {
     MergeJars() {
         output.convention(project.layout.buildDirectory.dir(name).map { it.file('output.jar') })
@@ -24,8 +26,8 @@ abstract class MergeJars extends DefaultTask {
         try (ZipOutputStream zout = new ZipOutputStream(new FileOutputStream(output.get().asFile))) {
             for (def jar : jars) {
                 try (ZipInputStream zin = new ZipInputStream(new FileInputStream(jar))) {
-                    def entry
-                    while ((entry = zin.getNextEntry()) != null) {
+                    ZipEntry entry
+                    while ((entry = zin.getNextEntry()) !== null) {
                         ZipEntry _new = new ZipEntry(entry.getName())
                         _new.setTime(0) //SHOULD be the same time as the main entry, but NOOOO _new.setTime(entry.getTime()) throws DateTimeException, so you get 0, screw you!
                         zout.putNextEntry(_new)