--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/ObjectTarget.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/ObjectTarget.groovy
@@ -27,6 +27,6 @@ public class ObjectTarget implements Comparable<ObjectTarget> {
     
     @Override
     int compareTo(ObjectTarget o) {
-        return toString().compareTo(o.toString())
+        return toString() <=> o.toString()
     }
 }