--- a/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/checks/CheckMode.groovy
+++ b/buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/checks/CheckMode.groovy
@@ -6,4 +6,6 @@ import groovy.transform.CompileStatic
 enum CheckMode {
     CHECK,
     FIX
+
+    CheckMode() {}
 }