--- a/fmlcore/build.gradle
+++ b/fmlcore/build.gradle
@@ -40,7 +40,7 @@ tasks.named('jar', Jar).configure {
     }
 }
 
-tasks.withType(JavaCompile) {
+tasks.withType(JavaCompile).configureEach {
     options.compilerArgs << '-Xlint:-unchecked'
 }