--- a/gradle.properties
+++ b/gradle.properties
@@ -2,6 +2,7 @@
 # This is required to provide enough memory for the Minecraft decompilation process.
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
+org.gradle.parallel=true
 
 JAVA_VERSION=17
 MC_VERSION=1.20.4