… CREDITS.txt
… buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/checks/CheckSAS.groovy
… buildSrc/src/main/groovy/net/minecraftforge/forge/tasks/checks/CheckTask.groovy
… build_forge.gradle
… build_forge_eclipse.gradle
… build_shared.gradle
… fmlearlydisplay/build.gradle
… fmlloader/build.gradle
… javafmllanguage/build.gradle
… lowcodelanguage/build.gradle
… mclanguage/build.gradle