@LexManos 2024-01-09
> Ordering of launcher json entries is no longer consistent

This should be a simple fix by using a LinkedHashMap for the json. Its kinda import that this is ordered for both reproducibility and for people reading the json.